# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TransJakarta Routes API
  description: |
    A RESTful API application built with Go (Gin framework) to manage TransJakarta BRT route information, stops/terminals, vehicles, and user reports.

    ## Features
    - User Authentication (Email/Password and OAuth)
    - Guest Access (read-only routes, stops, vehicles; create reports)
    - Role-Based Access Control (Common users and Admins)
    - Route Management
    - Stop/Terminal Management
    - Vehicle Management
    - Reporting System
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server
...
paths:
...
  /api/v1/auth/register:
    post:
      tags:
        - Authentication
      summary: Register new user
      description: Register a new user with email and password
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data:
                  id: 1
                  email: user@example.com
                  username: johndoe
                  role: common_user
                  created_at: "2024-01-01T00:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/auth/login:
    post:
      tags:
        - Authentication
      summary: Login with email/password
      description: Authenticate user and receive JWT tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/oauth/{provider}:
    get:
      tags:
        - Authentication
      summary: Initiate OAuth flow
      description: Redirect to OAuth provider (Google)
      operationId: oauthInitiate
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google]
      responses:
        "302":
          description: Redirect to OAuth provider
  /api/v1/auth/oauth/{provider}/callback:
    get:
      tags:
        - Authentication
      summary: OAuth callback
      description: Handle OAuth callback and authenticate user
      operationId: oauthCallback
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google]
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OAuth authentication successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/stops:
    get:
      tags:
        - Stops
      summary: List all stops
      description: Get paginated list of stops with optional filters. **Guest accessible** (authentication optional)
      operationId: listStops
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: type
          in: query
          schema:
            type: string
            enum: [stop, terminal]
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of stops
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StopListResponse"
    post:
      tags:
        - Stops
      summary: Create stop
      description: Create a new stop (Admin only). Supports both JSON-only and multipart/form-data with photo upload.
      operationId: createStop
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: JSON string of stop data
                  example: '{"name":"Halte Bundaran HI","type":"stop","latitude":-6.1944,"longitude":106.8229}'
                photo:
                  type: string
                  format: binary
                  description: "Photo file (optional). Allowed types: image/jpeg, image/png, image/webp, image/gif. Max size: 5MB"
            example:
              data: '{"name":"Halte Bundaran HI","type":"stop","latitude":-6.1944,"longitude":106.8229}'
              photo: ""
      responses:
        "201":
          description: Stop created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/stops/{id}:
    get:
      tags:
        - Stops
      summary: Get stop by ID
      description: Get detailed information about a specific stop. **Guest accessible** (authentication optional)
      operationId: getStop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stop details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Stops
      summary: Update stop
      description: Update stop information (Admin only). Supports both JSON-only and multipart/form-data with photo upload.
      operationId: updateStop
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                data:
                  type: string
                  description: JSON string of stop data
                  example: '{"name":"Halte Bundaran HI Updated","type":"stop","latitude":-6.1944,"longitude":106.8229}'
                photo:
                  type: string
                  format: binary
                  description: "Photo file (optional). Allowed types: image/jpeg, image/png, image/webp, image/gif. Max size: 5MB"
            example:
              data: '{"name":"Halte Bundaran HI Updated","type":"stop","latitude":-6.1944,"longitude":106.8229}'
              photo: ""
      responses:
        "200":
          description: Stop updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags:
        - Stops
      summary: Delete stop
      description: Delete a stop (Admin only)
      operationId: deleteStop
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stop deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/routes:
    get:
      tags:
        - Routes
      summary: List all routes
      description: Get paginated list of routes with optional filters. **Guest accessible** (authentication optional). Supports fuzzy search with typo tolerance.
      operationId: listRoutes
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive]
        - name: route_number
          in: query
          schema:
            type: string
        - name: search
          in: query
          description: Fuzzy search term to match against route name, route number, or description. Supports typo tolerance.
          schema:
            type: string
          example: "jakrta"
      responses:
        "200":
          description: List of routes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteListResponse"
    post:
      tags:
        - Routes
      summary: Create route
      description: Create a new route with stops (Admin only)
      operationId: createRoute
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRouteRequest"
      responses:
        "201":
          description: Route created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/routes/{id}:
    get:
      tags:
        - Routes
      summary: Get route by ID
      description: Get detailed information about a specific route with stops. **Guest accessible** (authentication optional)
      operationId: getRoute
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags:
        - Routes
      summary: Update route
      description: Update route metadata (Admin only)
      operationId: updateRoute
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Route"
      responses:
        "200":
          description: Route updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags:
        - Routes
      summary: Delete route
      description: Delete a route (Admin only)
      operationId: deleteRoute
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Route deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/routes/{id}/stops:
    put:
      tags:
        - Routes
      summary: Update route stops
      description: Update the stops for a route (Admin only)
      operationId: updateRouteStops
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRouteStopsRequest"
      responses:
        "200":
          description: Route stops updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/reports/{id}:
    get:
      tags:
        - Reports
      summary: Get report by ID
      description: Get detailed information about a specific report
      operationId: getReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Reports
      summary: Delete report
      description: Delete a report (Admin only)
      operationId: deleteReport
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Report deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    RegisterRequest:
      type: object
      required:
        - email
        - username
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          minLength: 3
          maxLength: 50
          example: johndoe
        password:
          type: string
          format: password
          minLength: 6
          example: password123
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          example: password123
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    MessageResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation completed successfully
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            success: false
            error: "Invalid request parameters"
...